refactor: Move `session-created`
