+++ frontend/src/app/sessions/user-sessions-wrapper/active-sessions/session-created/guacamole/guacamole.component.ts
